Updated --reorder also in merc-pbes.
