+++ tools/mcrl2/lps/src/main.rs
@@ -1,4 +1,3 @@
-use std::fmt;
 use std::fs::File;
 use std::io::BufWriter;
 use std::path::PathBuf;
@@ -18,11 +17,13 @@ use merc_lts::LtsStream;
 use merc_lts::MutexLtsBuilder;
 use merc_lts::guess_lts_output_format;
 use merc_symbolic::ExplorationStrategy as SymbolicExplorationStrategy;
+use merc_symbolic::Order;
 use merc_symbolic::ReachabilityOptions;
 use merc_symbolic::SummandGrouping;
 use merc_symbolic::SymbolicLTS;
 use merc_symbolic::SymbolicLpsOptions;
 use merc_symbolic::VariableOrder;
+use merc_symbolic::parse_order;
 use merc_symbolic::write_symbolic_lts;
 use merc_tools::KaHyParArgs;
 use merc_tools::VerbosityFlag;
@@ -102,54 +103,6 @@ enum Commands {
     ExploreExplicit(ExploreExplicitArgs),
 }
 
-/// The order in which process parameters are considered for reordering.
-#[derive(Debug, Clone)]
-enum Order {
-    /// Do not reorder the process parameters, the default.
-    None,
-
-    /// The MINCE algorithm for reordering process parameters, requires the
-    /// 'kahypar' tool.
-    Mince,
-
-    /// An explicit order given as a whitespace separated string of numbers.
-    Explicit(Vec<usize>),
-}
-
-impl fmt::Display for Order {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::None => write!(f, "none"),
-            Self::Mince => write!(f, "mince"),
-            Self::Explicit(order) => {
-                for (i, index) in order.iter().enumerate() {
-                    if i > 0 {
-                        write!(f, " ")?;
-                    }
-                    write!(f, "{index}")?;
-                }
-                Ok(())
-            }
-        }
-    }
-}
-
-fn parse_order(s: &str) -> Result<Order, String> {
-    match s.to_lowercase().as_str() {
-        "none" => Ok(Order::None),
-        "mince" => Ok(Order::Mince),
-        _ => {
-            // Parse the permutation
-            let permutation = s
-                .split_whitespace()
-                .map(|s| s.parse::<usize>())
-                .collect::<Result<Vec<_>, _>>()
-                .map_err(|error| format!("Failed to parse explicit order: {}", error.to_string()))?;
-            Ok(Order::Explicit(permutation))
-        }
-    }
-}
-
 /// The input LPS shared by every subcommand.
 #[derive(clap::Args, Debug)]
 struct InputArgs {
@@ -222,19 +175,7 @@ struct ExploreArgs {
 impl ExploreArgs {
     /// Returns the variable order to explore with, resolving the KaHyPar tool when `--reorder` is set.
     fn variable_order(&self) -> Result<VariableOrder, MercError> {
-        match &self.reorder {
-            Order::None => return Ok(VariableOrder::None),
-            Order::Mince => {
-                let (kahypar_path, kahypar_ini_path) = self.kahypar.resolve()?;
-                return Ok(VariableOrder::Mince {
-                    kahypar_path,
-                    kahypar_ini_path,
-                });
-            }
-            Order::Explicit(order) => {
-                return Ok(VariableOrder::Explicit(order.clone()));
-            }
-        }
+        self.reorder.resolve(|| self.kahypar.resolve())
     }
 }
