+++ tools/mcrl2/pbes/Cargo.toml
@@ -18,7 +18,7 @@ metrics = ["merc_pbes/metrics"]
 mcrl2.workspace = true
 merc_explore = { workspace = true, features = ["clap"] }
 merc_pbes.workspace = true
-merc_symbolic.workspace = true
+merc_symbolic = { workspace = true, features = ["clap"] }
 merc_tools.workspace = true
 merc_unsafety.workspace = true
 merc_utilities.workspace = true
